aarch64: Fix testcase pr112105.c
This testcase started to fail with r15-268-g9dbff9c05520a7.
When late_combine was added, it was turned on for -O2+ only,
so this testcase still failed.
This changes the option to be -O2 instead of -O and the testcase
started to pass again.

tested for aarch64-linux-gnu.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/pr112105.c: Change to be -O2 rather
	than -O1.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>