ada: Move Incomplete_View from node to entity field
The Incomplete_View property of a type was attached to its full type
declaration as a semantic field, but retrieving it from there required
low-level tree navigation and caused code duplication. In one case we
relied on internal class-wide type being attached to the corresponding
full type declaration, which is an undocumented assumption.

It seems better to attach this field to entities, just like we do with
Full_View and many other type properties. Ideally, this field should be
present just in type entities, but currently we set it before setting
the proper entity kind.

Behavior is unaffected. This is rather a code cleanup, originating from the
need to use Incomplete_View in GNATprove.

gcc/ada/ChangeLog:

	* einfo.ads (Incomplete_View): Move from Sinfo; adapt wording.
	* exp_ch3.adb (Build_Record_Init_Proc): Adapt retrieval of
	Incomplete_View.
	* gen_il-fields.ads (Opt_Field_Enum): Move Incomplete_View from node
	to entity field.
	* gen_il-gen-gen_entities.adb (Gen_Entities): Add field.
	* gen_il-gen-gen_nodes.adb (Gen_Nodes): Remove field.
	* sem_ch3.adb (Analyze_Full_Type_Declaration,
	Check_Anonymous_Access_Component): Adapt setting of Incomplete_View.
	* sem_ch6.adb (Analyze_Subprogram_Body_Helper): Adapt retrieval of
	Incomplete_View for class-wide types; no longer rely on class-wide
	type being attached to non-classwide type declaration.
	* sem_util.adb (Collect_Primitive_Operations): Adapt retrieval of
	Incomplete_View.
	* sinfo.ads (Incomplete_View): Move to Einfo.