testsuite: Fix sve/var_stride_*.c failures
gcc.target/aarch64/sve/var_stride_2.c started failing after
r15-268-g9dbff9c05520, but the change was an improvement:

@@ -36,13 +36,11 @@
 	b.any	.L9
 	ret
 .L17:
-	ubfiz	x5, x3, 10, 16
-	ubfiz	x4, x2, 10, 16
-	add	x5, x1, x5
-	add	x4, x0, x4
-	cmp	x0, x5
-	ccmp	x1, x4, 2, ls
 	uxtw	x4, w2
+	add	x6, x1, x3, lsl 10
+	cmp	x0, x6
+	add	x5, x0, x4, lsl 10
+	ccmp	x1, x5, 2, ls
 	ccmp	w2, 0, 4, hi
 	beq	.L3
 	cntb	x7

This patch therefore changes the test to expect the new output
for var_stride_2.c.

The changes for var_stride_4.c were a wash, with both versions
having 18(!) arithmetic instructions before the alias check branch.
Both versions sign-extend the n and m arguments as part of this
sequence; the question is whether they do it first or later.

This patch therefore changes the test to accept either the old
or the new code for var_stride_4.c.

gcc/testsuite/
	* gcc.target/aarch64/sve/var_stride_2.c: Expect ADD+LSL.
	* gcc.target/aarch64/sve/var_stride_4.c: Accept LSL or SBFIZ.