vect: Use original LHS type for gather pattern [PR118950].
In PR118950 we do not zero masked elements in a gather load.
While recognizing a gather/scatter pattern we do not use the original
type of the LHS.  This matters because the type can differ with bool
patterns (e.g. _Bool vs unsigned char) and we don't notice the need
for zeroing out the padding bytes.

This patch just uses the original LHS's type.

	PR middle-end/118950

gcc/ChangeLog:

	* tree-vect-patterns.cc (vect_recog_gather_scatter_pattern): Use
	original LHS's type.

gcc/testsuite/ChangeLog:

	* gcc.target/riscv/rvv/autovec/pr118950.c: New test.